+++ spec/integrations/cassandra_spec.rb
@@ -161,7 +161,7 @@ describe 'datadog::cassandra' do
     it { is_expected.to add_datadog_monitor('cassandra') }
 
     it 'renders expected YAML config file' do
-      expect(chef_run).to(render_file('/etc/datadog-agent/conf.d/cassandra.yaml').with_content { |content|
+      expect(chef_run).to(render_file('/etc/datadog-agent/conf.d/cassandra.d/conf.yaml').with_content { |content|
         expect(YAML.safe_load(content).to_json).to be_json_eql(YAML.safe_load(expected_yaml).to_json)
       })
     end
@@ -365,7 +365,7 @@ describe 'datadog::cassandra' do
     it { is_expected.to add_datadog_monitor('cassandra') }
 
     it 'renders expected YAML config file' do
-      expect(chef_run).to(render_file('/etc/datadog-agent/conf.d/cassandra.yaml').with_content { |content|
+      expect(chef_run).to(render_file('/etc/datadog-agent/conf.d/cassandra.d/conf.yaml').with_content { |content|
         expect(YAML.safe_load(content).to_json).to be_json_eql(YAML.safe_load(expected_yaml).to_json)
       })
     end
@@ -843,7 +843,7 @@ describe 'datadog::cassandra' do
     it { is_expected.to add_datadog_monitor('cassandra') }
 
     it 'renders expected YAML config file' do
-      expect(chef_run).to(render_file('/etc/datadog-agent/conf.d/cassandra.yaml').with_content { |content|
+      expect(chef_run).to(render_file('/etc/datadog-agent/conf.d/cassandra.d/conf.yaml').with_content { |content|
         expect(YAML.safe_load(content).to_json).to be_json_eql(YAML.safe_load(expected_yaml).to_json)
       })
     end