+++ spec/integrations/dns_check_spec.rb
@@ -56,7 +56,7 @@ describe 'datadog::dns_check' do
   it { is_expected.to add_datadog_monitor('dns_check') }
 
   it 'renders expected YAML config file' do
-    expect(chef_run).to(render_file('/etc/datadog-agent/conf.d/dns_check.yaml').with_content { |content|
+    expect(chef_run).to(render_file('/etc/datadog-agent/conf.d/dns_check.d/conf.yaml').with_content { |content|
       expect(YAML.safe_load(content).to_json).to be_json_eql(YAML.safe_load(expected_yaml).to_json)
     })
   end