+++ spec/integrations/elasticsearch_spec.rb
@@ -48,7 +48,7 @@ describe 'datadog::elasticsearch' do
   it { is_expected.to add_datadog_monitor('elastic') }
 
   it 'renders expected YAML config file' do
-    expect(chef_run).to(render_file('/etc/datadog-agent/conf.d/elastic.yaml').with_content { |content|
+    expect(chef_run).to(render_file('/etc/datadog-agent/conf.d/elastic.d/conf.yaml').with_content { |content|
       expect(YAML.safe_load(content).to_json).to be_json_eql(YAML.safe_load(expected_yaml).to_json)
     })
   end