+++ spec/integrations/go_expvar_spec.rb
@@ -57,7 +57,7 @@ describe 'datadog::go_expvar' do
   it { is_expected.to add_datadog_monitor('go_expvar') }
 
   it 'renders expected YAML config file' do
-    expect(chef_run).to(render_file('/etc/datadog-agent/conf.d/go_expvar.yaml').with_content { |content|
+    expect(chef_run).to(render_file('/etc/datadog-agent/conf.d/go_expvar.d/conf.yaml').with_content { |content|
       expect(YAML.safe_load(content).to_json).to be_json_eql(YAML.safe_load(expected_yaml).to_json)
     })
   end