+++ test/integration/datadog_jmx/serverspec/dd-agent-jmx_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-JMX_CONFIG = File.join(@agent_config_dir, 'conf.d/jmx.yaml')
+JMX_CONFIG = File.join(@agent_config_dir, 'conf.d/jmx.d/conf.yaml')
 
 describe service(@agent_service_name) do
   it { should be_running }