+++ test/integration/helpers/serverspec/spec_template.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-AGENT_CONFIG = File.join(@agent_config_dir, 'conf.d/REPLACEME.yaml')
+AGENT_CONFIG = File.join(@agent_config_dir, 'conf.d/REPLACEME.d/conf.yaml')
 
 describe service(@agent_service_name) do
   it { should be_running }